+++ plugins/guests/linux/cap/rsync.rb
@@ -18,8 +18,10 @@ module VagrantPlugins
 
         def self.rsync_post(machine, opts)
           machine.communicate.sudo(
-            "find '#{opts[:guestpath]}' '(' ! -user #{opts[:owner]} -or ! -group #{opts[:group]} ')' -print0 | " +
-            "xargs -0 -r chown -v #{opts[:owner]}:#{opts[:group]}")
+            "find '#{opts[:guestpath]}' " +
+            "'!' -type l -a " +
+            "'(' ! -user #{opts[:owner]} -or ! -group #{opts[:group]} ')' -print0 | " +
+            "xargs -0 -r chown #{opts[:owner]}:#{opts[:group]}")
         end
       end
     end